+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: jmanani <jmanani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/15 12:35:49 by jmanani           #+#    #+#             */
-/*   Updated: 2025/10/15 15:42:32 by jmanani          ###   ########.fr       */
+/*   Updated: 2025/10/15 15:43:57 by jmanani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int				ft_isascii(int inp);
 int				ft_isprint(int inp);
 size_t			ft_strlen(const char *s);
 void			*ft_memset(void *s, int c, size_t n);
-void			ft_bzero(void *s, int n);
+void			ft_bzero(void *s, size_t n);
 // memcpy
 void			*ft_memmove(void *dest, const void *src, int n);
 unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size);
